examples/docker-in-e2b/python/pyproject.toml
[tool.poetry]
name = "python-example"
version = "0.1.0"
description = ""
authors = ["Vasek Mlejnsky <vasek.mlejnsky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
e2b = "^1.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
